kits/python/pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arp-standard-server"
version = "0.3.7"
description = "FastAPI server scaffolding for the ARP Standard (v1)."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "Agent Runtime Protocol" }]
dependencies = [
  "fastapi>=0.110.0",
  "PyJWT[crypto]>=2.8.0",
  "arp-standard-model==0.3.7",
]

[project.urls]
Repository = "https://github.com/AgentRuntimeProtocol/ARP_Standard"
Issues = "https://github.com/AgentRuntimeProtocol/ARP_Standard/issues"

[project.optional-dependencies]
dev = ["pyright>=1.1.0", "pytest>=7"]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
arp_standard_server = ["py.typed"]
